+++ src/components/layout/IconMenu.tsx
@@ -8,7 +8,9 @@ import {
   StarFilled,
   UserOutlined,
 } from "@ant-design/icons";
+import { Image } from "antd";
 import React from "react";
+import { color, icon } from "../../resource";
 
 export const IconMenu = {
   task: (
@@ -44,11 +46,10 @@ export const IconMenu = {
     />
   ),
   campaign: (
-    <StarFilled
-      style={{
-        fontSize: "16px",
-        color: "#231F20",
-      }}
+    <Image
+      src={icon.pointRankInActive}
+      preview={false}
+      style={{ width: 18, height: 22, color: color.BK, paddingBottom: 4 }}
     />
   ),
   admin: (
@@ -68,11 +69,10 @@ export const IconMenu = {
     />
   ),
   mission: (
-    <RadarChartOutlined
-      style={{
-        fontSize: "16px",
-        color: "#231F20",
-      }}
+    <Image
+      src={icon.targetInActive}
+      preview={false}
+      style={{ width: 18, height: 18, color: color.BK }}
     />
   ),
 };
@@ -110,11 +110,10 @@ export const IconMenuInActive = {
     />
   ),
   campaign: (
-    <StarFilled
-      style={{
-        fontSize: "16px",
-        color: "#FFCA37",
-      }}
+    <Image
+      src={icon.pointRankActive}
+      preview={false}
+      style={{ width: 18, height: 22, color: color.BK, paddingBottom: 4 }}
     />
   ),
   admin: (
@@ -134,11 +133,10 @@ export const IconMenuInActive = {
     />
   ),
   mission: (
-    <RadarChartOutlined
-      style={{
-        fontSize: "16px",
-        color: "#FFCA37",
-      }}
+    <Image
+      src={icon.targetActive}
+      preview={false}
+      style={{ width: 18, height: 18 }}
     />
   ),
 };